Add RestUserCommunity
This change aims to make community entities handle different users a
little more gracefully. To make changes (or get some kinds of
information) for a community you simply create a UserCommunity from the
available method found on User objects (if you have a user and not a
community) or Community objects (if you have a community and not a
user).